Updated alarm logic, (hopefully) finalised add functionality
implemented duplicate testing to prevent alarms from being set to the same time

preliminary logic for the alarm triggering added, alongside replacing PriorityQueue's Person with a new AlarmTime class that holds the display text taken from the date picker and a corresponding datetime for scheduling the alarm

Add prevention against alarms being set in the past

Reimplemented PriorityQueue's SortedArray as a part of ClockV2 in order to switch to using a sorted queue and to reverse the add order as using epoch times requires priority goes from lowest to highest.
Originally planned to extended PriorityQueue's SortedArray but as it's variables are readonly and I want to leave it unmodified, the easiest solution was a full reimplementation.

impleted add alarm in the subform, uses epoch time for priority

testing adding an alarm queue and displaying its top item